moved data_management to the root, created flatten method
