+++ cpp/onedal/array.hpp
@@ -14,34 +14,32 @@
  * limitations under the License.
  *******************************************************************************/
 
-#include "onedal/data_management/array.hpp"
-#include "onedal/data_management/detail/array_impl.hpp"
+#pragma once
 
-using std::int32_t;
-using std::int64_t;
+#include "onedal/common.hpp"
+#include "onedal/detail/common.hpp"
 
 namespace dal {
-namespace data_management {
+
+namespace detail {
+template <typename T> class array_impl;
+} // namespace detail
 
 template <typename T>
-T* array<T>::get_data() noexcept {
-    return _impl->get_data_ptr();
-}
+class array {
+public:
+    using pimpl = dal::detail::pimpl< detail::array_impl<T> >;
 
-template <typename T>
-const T* array<T>::get_data() const noexcept {
-    const auto& arr_impl = *_impl;
-    return arr_impl.get_data_ptr();
-}
+public:
+    array(const pimpl& impl)
+        : _impl(impl)
+    { }
 
-template <typename T>
-int64_t array<T>::get_size() const noexcept {
-    return _impl->get_size();
-}
+    T* get_pointer() const noexcept;
+    std::int64_t get_size() const noexcept;
 
-template class array<float>;
-template class array<double>;
-template class array<int32_t>;
+private:
+    pimpl _impl;
+};
 
-} // namespace data_management
 } // namespace dal
