--- cpp/onedal/data_management/BUILD.bazel
@@ -1,19 +1,0 @@
-load("//build/bazel:rules.bzl", "dal_module")
-
-filegroup(
-  name = "private_headers",
-  srcs = glob(["detail/*.hpp"]),
-)
-
-filegroup(
-  name = "public_headers",
-  srcs = glob(["**/*.hpp"], exclude=[":private_headers"]),
-)
-
-dal_module(
-    name = "data_management",
-    srcs = glob(["**/*.cpp"]) + [":private_headers"],
-    hdrs = [":public_headers"],
-    deps = ["//cpp/onedal:common"],
-    visibility = ["//visibility:public"],
-)
