--- cpp/onedal/data_management/detail/table_impl.hpp
@@ -1,44 +1,0 @@
-/*******************************************************************************
- * Copyright 2020 Intel Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *******************************************************************************/
-
-#pragma once
-
-#include "onedal/common.hpp"
-#include "onedal/data_management/detail/table_data.hpp"
-
-namespace dal {
-namespace data_management {
-namespace detail {
-
-class table_impl {
-public:
-    table_data_ptr data_container;
-    slice elements_to_access;
-
-    table_impl(const table_data_ptr& cnt, const slice& s)
-        : data_container(cnt),
-          elements_to_access(s)
-    { }
-
-    table_impl(const table_data_ptr& cnt)
-        : data_container(cnt),
-          elements_to_access({ .rows = {0, -1}, .cols = {0, -1} })
-    { }
-};
-
-} // namespace detail
-} // namespace data_management
-} // namespace dal
