+++ cpp/onedal/decomposition/pca/train_types.hpp
@@ -29,17 +29,17 @@ class train_result_impl;
 
 class train_input : public base {
  public:
-  train_input(const data_management::table& data);
+  train_input(const table& data);
 
-  data_management::table get_data() const;
+  table get_data() const;
 
-  auto& set_data(const data_management::table& data) {
+  auto& set_data(const table& data) {
     set_data_impl(data);
     return *this;
   }
 
  private:
-  void set_data_impl(const data_management::table& data);
+  void set_data_impl(const table& data);
   dal::detail::pimpl<detail::train_input_impl> impl_;
 };
 
@@ -51,9 +51,9 @@ class train_result {
   train_result(const pimpl& impl);
 
   auto get_model() const -> model;
-  data_management::table get_eigenvalues() const;
-  data_management::table get_eigenvectors() const;
-  data_management::table get_explained_variance() const;
+  table get_eigenvalues() const;
+  table get_eigenvectors() const;
+  table get_explained_variance() const;
 
  private:
   dal::detail::pimpl<detail::train_result_impl> impl_;
