+++ cpp/onedal/homogen_table.hpp
@@ -16,19 +16,33 @@
 
 #pragma once
 
-#include "onedal/integral_types.hpp"
-#include "onedal/data_management/homogen_table.hpp"
+#include "onedal/table.hpp"
 
 namespace dal {
-namespace data_management {
 
 namespace detail {
+class homogen_table_data;
+} // namespace detail
+
+class homogen_table : public table {
+public:
+    using pimpl = dal::detail::pimpl<detail::homogen_table_data>;
+
+    homogen_table(const homogen_table& t)
+        : table(t.get_impl())
+    { }
+
+    template<typename DataType>
+    homogen_table(const DataType* data, std::int64_t rows, std::int64_t cols, data_format df);
+
+    detail::homogen_table_data* get_impl_ptr() const noexcept {
+        return reinterpret_cast<detail::homogen_table_data*>(table::get_impl_ptr());
+    }
+};
 
 template <typename DataType>
 homogen_table create_table(const DataType* data,
                            std::int64_t rows, std::int64_t cols,
                            data_format df = data_format::rowmajor);
 
-} // namespace detail
-} // namespace data_management
 } // namespace dal
