+++ examples/cpp/onedal/standalone/BUILD
@@ -6,7 +6,6 @@ cc_binary(
   copts = COMMON_COPTS,
   deps = [
     "//cpp/onedal/decomposition/pca",
-    "//cpp/onedal/data_management",
     "//cpp/daal:threading_tbb",
   ],
 )
